moving HomeScreen.js from components directory to screens directory
