+++ frontend/src/App.js
@@ -2,7 +2,7 @@ import Header from './components/Header'
 import Footer from './components/Footer'
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 import { Container } from 'react-bootstrap'
-import HomeScreen from './components/HomeScreen'
+import HomeScreen from './screens/HomeScreen'
 import ProductScreen from './screens/ProductScreen'
 
 const App = () => {
