+++ frontend/src/screens/HomeScreen.js
@@ -1,6 +1,6 @@
 import { Row, Col } from "react-bootstrap"
 import products from "../products"
-import Product from './Product'
+import Product from '../components/Product'
 
 const HomeScreen = () => {
     return (
